Keep fixtures up to date with registry changes; Deprecate or relocate some test utilities.
