+++ tests/blockchain/eth/utilities.py
@@ -1,15 +1,11 @@
 import random
 from typing import List
 
-import pkg_resources
 from constant_sorrow import constants
-from eth_tester import PyEVMBackend
-from eth_tester.backends import is_pyevm_available
-from eth_tester.backends.pyevm.main import get_default_genesis_params, get_default_account_keys, generate_genesis_state
 from umbral import keys
 from web3 import Web3
 
-from nucypher.blockchain.eth.agents import MinerAgent, NucypherTokenAgent
+from nucypher.blockchain.eth.agents import MinerAgent
 from nucypher.blockchain.eth.deployers import MinerEscrowDeployer, NucypherTokenDeployer
 
 
@@ -73,64 +69,3 @@ def generate_accounts(w3: Web3, quantity: int) -> List[str]:
         w3.personal.unlockAccount(address, passphrase=insecure_passphrase)
         addresses.append(addresses)
     return addresses
-
-
-class TesterPyEVMBackend(PyEVMBackend):
-
-    def __init__(self, genesis_overrides=None):
-        """
-        Example overrides
-        ---------------------------------
-        coinbase: address_bytes
-        difficulty: int(default: 131072)
-        extra_data: bytes
-        gas_limit: int(default: 3141592)
-        gas_used: int(default: 0)
-        nonce: bytes
-        block_number: int
-        timestamp: int(epoch)
-        """
-
-        self.fork_config = dict()
-
-        if not is_pyevm_available():
-            raise pkg_resources.DistributionNotFound(
-                "The `py-evm` package is not available.  The "
-                "`PyEVMBackend` requires py-evm to be installed and importable. "
-                "Please install the `py-evm` library."
-            )
-
-        self.reset_to_genesis(overrides=genesis_overrides)
-
-    def reset_to_genesis(self, overrides=None):
-        from evm.chains.tester import MainnetTesterChain
-        from evm.db import get_db_backend
-
-        base_db = get_db_backend()
-        genesis_params = get_default_genesis_params()
-
-        # Genesis params overrides
-        if overrides is not None:
-            genesis_params.update(overrides)
-
-        account_keys = get_default_account_keys()
-        genesis_state = generate_genesis_state(account_keys)
-        chain = MainnetTesterChain.from_genesis(base_db, genesis_params, genesis_state)
-
-        self.account_keys, self.chain = account_keys, chain
-
-
-def token_airdrop(token_agent, amount: int, origin: str, addresses: List[str]):
-    """Airdrops tokens from creator address to all other addresses!"""
-
-    def txs():
-        for address in addresses:
-            txhash = token_agent.contract.functions.transfer(address, amount).transact({'from': origin,
-                                                                                 'gas': 2000000})
-            yield txhash
-
-    receipts = list()
-    for tx in txs():    # One at a time
-        receipt = token_agent.blockchain.wait_for_receipt(tx)
-        receipts.append(receipt)
-    return receipts
